// auth/features/ory-elements/types.ts
export type OryNodeInputInputProps = {
  name: string
  value: string | number | readonly string[] | undefined
  onClick: MouseEventHandler
  /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
  onChange: (event: any) => void

  onBlur: () => void
  /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
  ref: (instance: any) => void
  disabled?: boolean
  type: string
  autoComplete?: string
  maxLength?: number
  placeholder: string
}

export type OryNodeInputProps = {
  /** @deprecated - use node.attributes */
  attributes: UiNodeInputAttributes
  node: UiNodeInput
  /** @deprecated - use inputProps.onClick */
  onClick?: MouseEventHandler
  inputProps: OryNodeInputInputProps
}